requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart
pandas
pytest
httpx
pytest-asyncio